made fixed changed suggested by aecio
